fix: resolve sky-executive detail 404 by fetching data directly in getStaticProps
getStaticProps fetched a relative URL (/api/sky/executives/{id}), which
throws in Node since there is no origin to resolve it against. The error
was swallowed, leaving executive null and returning notFound for every
sky-executive detail page.

Extract the upstream fetch into a shared fetchSkyExecutiveDetail function
and call it directly from getStaticProps. The API route now imports the
same function (behavior unchanged); the client-side fetch keeps its
relative URL, which is correct in the browser.
